requirements.txt
faiss-cpu
sentence-transformers
langchain
langchain-text-splitters
google-generativeai
numpy
streamlit
